requirements.txt
babel==2.17.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
docopt==0.6.2
docx-mailmerge==0.5.0
docx2pdf==0.1.8
docxcompose==1.4.0
docxtpl==0.20.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
num2words==0.5.14
packaging==25.0
pyodbc==5.2.0
python-docx==1.2.0
python-dotenv==1.1.1
pywin32==311
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.42
tqdm==4.67.1
typing_extensions==4.14.1
waitress==3.0.2
Werkzeug==3.1.3
WTForms==3.2.1
WTForms-SQLAlchemy==0.4.2
requests==2.32.5
google-generativeai
pdfplumber 
easyocr
PyMuPDF 
numpy
pillow
mammoth